% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderator_plot.R
\name{moderator_plot}
\alias{moderator_plot}
\title{Moderator plot}
\usage{
moderator_plot(mod, x, dv, xlab, ylab, ref, foc)
}
\arguments{
\item{mod}{the model}

\item{x}{the moderator}

\item{dv}{the dependent variable}

\item{xlab}{the x-axis label}

\item{ylab}{the y-axis lable}

\item{ref}{the reference group}

\item{foc}{the focal group}
}
\description{
Create moderator plot for two contrasts (i.e, PH vs. HW, PH vs. ctrl, and HW vs. ctrl)
}
