% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_plot.R
\name{prob_plot}
\alias{prob_plot}
\title{Makes univariate probability plots}
\usage{
prob_plot(x, outcome, lab)
}
\arguments{
\item{x}{is the independent variable}

\item{outcome}{the dependent variable}

\item{lab}{is the x-axis label}
}
\description{
Makes univariate probability plots
}
