% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_plot.R
\name{rs_plot}
\alias{rs_plot}
\title{Regions of significance plot
Plots the regions of significance}
\usage{
rs_plot(mod, predictor, moderator, lb, ub, xlab, ylab, plot.jn = FALSE)
}
\arguments{
\item{mod}{The fitted model}

\item{predictor}{The name of focal predictor variable}

\item{moderator}{The name of the moderator}

\item{lb}{The lower bound for the moderator}

\item{ub}{The upper bound for the moderator}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}

\item{plot.jn}{Logical. Shoud we plot the points when the interaction is significantly different than zero?}
}
\description{
Regions of significance plot
Plots the regions of significance
}
