% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_plot.R
\name{prob_plot}
\alias{prob_plot}
\title{Makes univariate probability plots}
\usage{
prob_plot(mod, x, outcome, xlab, ylab)
}
\arguments{
\item{mod}{the model}

\item{x}{is the independent variable}

\item{outcome}{the dependent variable}

\item{xlab}{is the x-axis label}

\item{ylab}{is the x-axis label}
}
\description{
Makes univariate probability plots
}
